tests: Move test framework into its own crate
