+++ tests/Cargo.toml
@@ -7,39 +7,20 @@ license.workspace = true
 repository.workspace = true
 version.workspace = true
 
-[dependencies]
-futures = "0.3.29"
-ruffle_core = { path = "../core", features = ["deterministic", "timeline_debug", "avm_debug", "audio", "mp3", "default_font"] }
-ruffle_render_wgpu = { path = "../render/wgpu" }
-ruffle_render = { path = "../render" }
-ruffle_input_format = { path = "input-format" }
-ruffle_socket_format = { path = "socket-format" }
-ruffle_video_software = { path = "../video/software", optional = true }
-image = { version = "0.24.7", default-features = false, features = ["png"] }
-regex = "1.10.2"
-url = "2.4.1"
-chrono = "0.4.31"
-
 [features]
 # Enable running image comparison tests. This is off by default,
 # since the images we compare against are generated on CI, and may
 # not match your local machine's Vulkan version / image output.
-imgtests = ["ruffle_video_software"]
-jpegxr = ["ruffle_core/jpegxr"]
-lzma = ["ruffle_core/lzma"]
+imgtests = ["ruffle_test_framework/imgtests"]
+jpegxr = ["ruffle_test_framework/jpegxr"]
+lzma = ["ruffle_test_framework/lzma"]
 
 [dev-dependencies]
-approx = "0.5.1"
-pretty_assertions = "1.4.0"
-env_logger = "0.10.0"
-tracing = { workspace = true}
-tracing-subscriber = { workspace = true }
-serde = "1.0.190"
-toml = "0.8.6"
+ruffle_core = { path = "../core", features = ["deterministic", "timeline_debug", "avm_debug", "audio", "mp3", "default_font"] }
+ruffle_test_framework = { path = "framework" }
 libtest-mimic = "0.6.1"
 walkdir = "2.4.0"
 anyhow = "1.0.75"
-async-channel = "1.9.0"
 
 [[test]]
 name = "tests"
