+++ tests/framework/src/image_trigger.rs
