+++ tests/tests/regression_tests.rs
@@ -7,28 +7,12 @@ use crate::shared_object::{shared_object_avm1, shared_object_avm2, shared_object
 use anyhow::Context;
 use anyhow::Result;
 use libtest_mimic::{Arguments, Trial};
+use ruffle_test_framework::test::Test;
 use std::panic::{catch_unwind, resume_unwind};
 use std::path::Path;
-use util::test::Test;
 
 mod external_interface;
 mod shared_object;
-mod util;
-
-fn set_logger() {
-    let _ = env_logger::Builder::from_env(
-        env_logger::Env::default().default_filter_or("info,wgpu_core=warn,wgpu_hal=warn"),
-    )
-    .format_timestamp(None)
-    .is_test(true)
-    .try_init();
-
-    let subscriber = tracing_subscriber::fmt::Subscriber::builder()
-        .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
-        .finish();
-    // Ignore error if it's already been set
-    let _ = tracing::subscriber::set_global_default(subscriber);
-}
 
 fn is_candidate(args: &Arguments, test_name: &str) -> bool {
     if let Some(filter) = &args.filter {
